Module neli::attr

Shared attribute code for all types of netlink attributes.

This module is relatively small right now and will eventually contain more code once type parameters in associated types defined in traits are stabilized. Due to neli being supported on stable and nightly, I cannot currently use this feature and have opted to define implementations of the trait separately for Rtattr and Nlattr types in the rtnl.rs and genl.rs modules respectively.

Enums

AttrHandle

Handle returned for traversing nested attribute structures

AttrHandleMut

Handle for traversing nested attribute structures mutably

Traits

Attribute

Trait that defines shared operations for netlink attributes. Currently, this applies to generic netlink and routing netlink attributes.