Trait neli::attr::Attribute

pub trait Attribute<T> {
    fn payload(&self) -> &Buffer;
    fn set_payload<P>(&mut self, payload: &P) -> Result<(), NlError>
    where
        P: Nl;

    fn get_payload_as<R>(&self) -> Result<R, NlError>
    where
        R: Nl,
    { ... }
}

Trait that defines shared operations for netlink attributes. Currently, this applies to generic netlink and routing netlink attributes.

Required Methods

fn payload(&self) -> &Buffer

Get the payload of the given attribute.

Due to Rust’s requirement that all elements of a Vec are of the same type, payloads are represented as a byte buffer so that nested attributes that contain multiple types for the payload can be type checked before serialization yet still contained all in the same top level attribute.

fn set_payload<P>(&mut self, payload: &P) -> Result<(), NlError>where
P: Nl,

Set the payload to a data type that implements Nl - this function will overwrite the current payload.

This method serializes the payload parameter and stores the resulting byte buffer as the payload.

Provided Methods

fn get_payload_as<R>(&self) -> Result<R, NlError>where
R: Nl,

Get an Nlattr payload as the provided type parameter, R.