Module neli::consts::genl

Constants related to generic netlink

Structs

Index

Type representing attribute list types as indices

Enums

CtrlAttr

Values for nla_type in Nlattr

CtrlAttrMcastGrp

Values for nla_type in Nlattr

CtrlCmd

Values for cmd in Genlmsghdr.

NlAttrTypeWrapper

Wrapper that is usable with all values in the Nlattr field, nla_type.

Traits

Cmd

Trait marking constants valid for use in Genlmsghdr field, cmd.

NlAttrType

Marker trait for types usable in the Nlattr field, nla_type