Enum neli::consts::netfilter::LogCmd

pub enum LogCmd {
    Bind,
    Unbind,
    PfBind,
    PfUnbind,
    UnrecognizedConst(u8),
}

Command value for the NfLogCfg::Cmd.

Variants

Bind

Unbind

PfBind

PfUnbind

UnrecognizedConst(u8)

Implementations

impl LogCmd

pub fn is_unrecognized(&self) -> bool

Check whether a given method is an unrecognized constant for the set of possible constants associated with the current type.