neli::consts::nl
Enum NlmF

pub enum NlmF {

    Request,
    Multi,
    Ack,
    Echo,
    DumpIntr,
    DumpFiltered,
    Root,
    Match,
    Atomic,
    Dump,
    Replace,
    Excl,
    Create,
    Append,
    UnrecognizedConst(u16),

}

Values for nl_flags in Nlmsghdr

Variants

Request

This flag is required for all kernel requests

Multi

Ack

Echo

DumpIntr

DumpFiltered

Root

Match

Atomic

Dump

Replace

Excl

Create

Append

UnrecognizedConst(u16)

Implementations

impl NlmF

pub fn is_unrecognized(&self) -> bool

Check whether a given method is an unrecognized constant for the set of possible constants associated with the current type.