Module neli::consts::rtnl

Constants related to rtnetlink

Structs

IfaFFlags

IffFlags

NtfFlags

NudFlags

RtmFFlags

Enums

Af

Internet address families

Arphrd

Interface types

Ifa

Enum usable with Rtattr field, rta_type. Values are interface address message attributes. Used with Ifaddrmsg.

IfaF

Interface address flags

Iff

Values for ifi_flags in Ifinfomsg.

Ifla

Enum usable with Rtattr field, rta_type. Values are interface information message attributes. Used with Ifinfomsg.

IflaInfo

Enum usable with Rtattr field, rta_type. Values are nested attributes to IFLA_LINKMODE.

Nda

Enum usable with Rtattr field, rta_type. Values specify neighbor table attributes

Ntf

Arp neighbor cache entry flags

Nud

Arp neighbor cache entry states

RtAddrFamily

General address families for sockets

RtScope

rtm_scope The distance between destinations

RtTable

rt_class_t Reserved route table identifiers

Rta

Enum usable with Rtattr field, rta_type. Values are routing message attributes. Used with Rtmsg.

RtaTypeWrapper

Wrapper that is usable for all values in Rtattr field, rta_type

Rtm

rtnetlink-related values for nl_type in Nlmsghdr.

RtmF

rtm_flags Flags for rtnetlink messages

Rtn

rtm_type The results of a lookup from a route table

Rtprot

rtm_protocol The origins of routes that are defined in the kernel

Tca

Enum usable with Rtattr field, rta_type. Values specify queuing discipline attributes. Used with Tcmsg.

Traits

RtaType

Marker trait for Rtattr field, rta_type.