Enum neli::consts::rtnl::Ifla

pub enum Ifla {

    Unspec,
    Address,
    Broadcast,
    Ifname,
    Mtu,
    Link,
    Qdisc,
    Stats,
    Cost,
    Priority,
    Master,
    Wireless,
    Protinfo,
    Txqlen,
    Map,
    Weight,
    Operstate,
    Linkmode,
    Linkinfo,
    NetNsPid,
    Ifalias,
    NumVf,
    VfinfoList,
    Stats64,
    VfPorts,
    PortSelf,
    AfSpec,
    Group,
    NetNsFd,
    ExtMask,
    Promiscuity,
    NumTxQueues,
    NumRxQueues,
    Carrier,
    PhysPortId,
    CarrierChanges,
    PhysSwitchId,
    LinkNetnsid,
    PhysPortName,
    ProtoDown,
    GsoMaxSegs,
    GsoMaxSize,
    Pad,
    Xdp,
    Event,
    NewNetnsid,
    IfNetnsid,
    CarrierUpCount,
    CarrierDownCount,
    NewIfindex,
    MinMtu,
    MaxMtu,
    PropList,
    AltIfname,
    PermAddress,
    ProtoDownReason,
    UnrecognizedConst(c_ushort),

}

Enum usable with Rtattr field, rta_type. Values are interface information message attributes. Used with Ifinfomsg.

Variants

Unspec

Address

Broadcast

Ifname

Mtu

Link

Qdisc

Stats

Cost

Priority

Master

Wireless

Protinfo

Txqlen

Map

Weight

Operstate

Linkmode

Linkinfo

NetNsPid

Ifalias

NumVf

VfinfoList

Stats64

VfPorts

PortSelf

AfSpec

Group

NetNsFd

ExtMask

Promiscuity

NumTxQueues

NumRxQueues

Carrier

PhysPortId

CarrierChanges

PhysSwitchId

LinkNetnsid

PhysPortName

ProtoDown

GsoMaxSegs

GsoMaxSize

Pad

Xdp

Event

NewNetnsid

IfNetnsid

CarrierUpCount

CarrierDownCount

NewIfindex

MinMtu

MaxMtu

PropList

AltIfname

PermAddress

ProtoDownReason

UnrecognizedConst(c_ushort)

Implementations

impl Ifla

pub fn is_unrecognized(&self) -> bool

Check whether a given method is an unrecognized constant for the set of possible constants associated with the current type.