neli::consts::rtnl

Enum Rtn

pub enum Rtn {

    Unspec,
    Unicast,
    Local,
    Broadcast,
    Anycast,
    Multicast,
    Blackhole,
    Unreachable,
    Prohibit,
    Throw,
    Nat,
    Xresolve,
    UnrecognizedConst(c_uchar),

}

rtm_type The results of a lookup from a route table

Variants

Unspec

Unicast

Local

Broadcast

Anycast

Multicast

Blackhole

Unreachable

Prohibit

Throw

Nat

Xresolve

UnrecognizedConst(c_uchar)

Implementations

impl Rtn

pub fn is_unrecognized(&self) -> bool

Check whether a given method is an unrecognized constant for the set of possible constants associated with the current type.