Module neli::err

This is the module that contains the error types used in neli

There are four main types:

Nlmsgerr - an application error returned from netlink as a packet.

NlError - a general netlink error wrapping application errors, serialization and deserialization errors, and other errors that occur in neli.

DeError - error while deserializing

SerError - error while serializing

Design decisions

All errors implement std::error::Error in an attempt to allow them to be used in conjunction with Result for easier error management even at the protocol error level.

Deserializing the NlmsghdrErr struct simply consumes the remaining data available in the deserialization buffer for the nl_payload field. This is due to the fact that ACKs do not contain the payload of the packet that triggered the ACK, but errors do contain the payload of the packet that triggered the error. This means that the nl_len field of NlmsghdrErr cannot be relied on for deserialization. To access the payload of the packet that caused the error, use NlmsghdrErr::get_payload_as. See the examples/ directory for a usage example.

Structs

Nlmsgerr

Struct representing netlink packets containing errors

NlmsghdrErr

An Nlmsghdr header with no payload returned as part of errors.

Enums

DeError

Deserialization error

NlError

General netlink error

SerError

Serialization error

WrappedError

An error to wrap all system level errors in a single, higher level error.