neli::err
Struct NlmsghdrErr

pub struct NlmsghdrErr<T, P> {
    pub nl_len: u32,
    pub nl_type: T,
    pub nl_flags: NlmFFlags,
    pub nl_seq: u32,
    pub nl_pid: u32,
    pub nl_payload: P,
}

A special struct that represents the contents of an error returned at the application level. Because the returned nl_len cannot always determine the length of the packet (as in the case of ACKs where no payload will be returned), this data structure relies on the total packet size for deserialization.

Fields
nl_len: u32

Length of the netlink message

nl_type: T

Type of the netlink message

nl_flags: NlmFFlags

Flags indicating properties of the request or response

nl_seq: u32

Sequence number for netlink protocol

nl_pid: u32

ID of the netlink destination for requests and source for responses.

nl_payload: P

Payload of netlink message