Struct neli::genl::Nlattr

pub struct Nlattr<T, P> {
    pub nla_len: u16,
    pub nla_type: AttrType<T>,
    pub nla_payload: P,
}

Struct representing netlink attributes and payloads

Fields
nla_len: u16

Length of the attribute header and payload together

nla_type: AttrType<T>

Type information for the netlink attribute

nla_payload: P

Payload of the attribute - either parsed or a binary buffer

Implementations

impl<T> Nlattr<T, Buffer>
where T: NlAttrType,

pub fn new<P>( nla_nested: bool, nla_network_order: bool, nla_type: T, nla_payload: P ) -> Result<Self, SerError>
where P: Size + ToBytes,

Create a new Nlattr with parameters for setting bitflags in the header.

pub fn add_nested_attribute<TT, P>( &mut self, attr: &Nlattr<TT, P> ) -> Result<(), SerError>
where TT: NlAttrType, P: ToBytes,

Add a nested attribute to the end of the payload.

pub fn get_attr_handle<R>( &self ) -> Result<AttrHandle<'_, GenlBuffer<R, Buffer>, Nlattr<R, Buffer>>, DeError>
where R: NlAttrType,

Return an AttrHandle for attributes nested in the given attribute payload

pub fn get_attr_handle_mut<R>( &mut self ) -> Result<AttrHandleMut<'_, GenlBuffer<R, Buffer>, Nlattr<R, Buffer>>, DeError>
where R: NlAttrType,

Return a mutable AttrHandle for attributes nested in the given attribute payload