neli

Module nl

This module contains the top level netlink header code. Every netlink message will be encapsulated in a top level Nlmsghdr.

Nlmsghdr is the structure representing a header that all netlink protocols require to be passed to the correct destination.

Design decisions

Payloads for Nlmsghdr can be any type.

The payload is wrapped in an enum to facilitate better application-level error handling.

Structs

Nlmsghdr

Top level netlink header and payload

Enums

NlPayload

An enum representing either the desired payload as requested by the payload type parameter, an ACK received at the end of a message or stream of messages, or an error.