neli::nl

Struct Nlmsghdr

pub struct Nlmsghdr<T, P> { /* private fields */ }

Top level netlink header and payload

Implementations

impl<T, P> Nlmsghdr<T, P>

pub fn nl_len(&self) -> &u32

Length of the netlink message

pub fn nl_type(&self) -> &T

Type of the netlink message

pub fn nl_flags(&self) -> &NlmF

Flags indicating properties of the request or response

pub fn nl_seq(&self) -> &u32

Sequence number for netlink protocol

pub fn nl_pid(&self) -> &u32

ID of the netlink destination for requests and source for responses.

pub fn nl_payload(&self) -> &NlPayload<T, P>

Payload of netlink message

impl<T, P> Nlmsghdr<T, P>
where T: NlType,

pub fn get_payload(&self) -> Option<&P>

Get the payload if there is one or return an error.

pub fn get_err(&mut self) -> Option<Nlmsgerr<NlmsghdrErr<T, P>>>

Get an error from the payload if it exists.

Takes a mutable reference because the payload will be swapped for Empty to gain ownership of the error.

impl<T, P> Nlmsghdr<T, P>
where T: NlType, P: Size,

pub fn set_payload(&mut self, p: NlPayload<T, P>)

Set the payload for Nlmsghdr and handle the change in length internally.

impl Nlmsghdr<u16, Buffer>

pub fn to_typed<T, P>(self) -> Result<Nlmsghdr<T, P>, RouterError<T, P>>
where T: NlType, P: Size + FromBytesWithInput<Input = usize>,

Set the payload for Nlmsghdr and handle the change in length internally.