Enum neli::nlhdr::AttrHandle

pub enum AttrHandle<P> {
    Bin(Vec<u8>),
    Parsed(Vec<NlAttrHdr<P>>),
}

Handle returned by GenlHdr for traversing nested attribute structures

Variants
Bin(Vec<u8>)

Binary internal representation of attributes

Parsed(Vec<NlAttrHdr<P>>)

Rust representation of attributes

Methods

impl<'a, P> AttrHandle<P> where
P: PartialEq + Nl,

pub fn parse_nested_attributes(&mut self) -> Result<&mut AttrHandle<P>, DeError>

Parse a binary payload into nested attributes

pub fn get_nested_attributes<S>(
&mut self,
payload: P
) -> Result<AttrHandle<S>, DeError>

Get the payload of an attribute as a handle for parsing nested attributes

pub fn get_attribute(&'a self, p: P) -> Option<&'a NlAttrHdr<P>>

Get nested attributes from a parsed handle

pub fn get_attribute_mut(&'a mut self, p: P) -> Option<&'a mut NlAttrHdr<P>>

Mutably get nested attributes from a parsed handle

pub fn get_payload_as<T>(&mut self, attr: P) -> Result<T, DeError> where
T: Nl,

Parse binary payload as a type that implements Nl