Struct neli::nlhdr::NlHdr

pub struct NlHdr<I, T> {
    pub nl_len: u32,
    pub nl_type: I,
    pub nl_flags: Vec<NlFlags>,
    pub nl_seq: u32,
    pub nl_pid: u32,
    pub nl_payload: T,
}

Top level netlink header and payload

Fields
nl_len: u32

Length of the netlink message

nl_type: I

Type of the netlink message

nl_flags: Vec<NlFlags>

Flags indicating properties of the request or response

nl_seq: u32

Sequence number for netlink protocol

nl_pid: u32

ID of the netlink destination for requests and source for responses

nl_payload: T

Payload of netlink message

Methods

impl<I: Nl, T: Nl> NlHdr<I, T>

pub fn new(
nl_len: Option<u32>,
nl_type: I,
nl_flags: Vec<NlFlags>,
nl_seq: Option<u32>,
nl_pid: Option<u32>,
nl_payload: T
) -> Self

Create a new top level netlink packet with a payload