neli::router::synchronous

Struct NlRouter

pub struct NlRouter { /* private fields */ }

A high-level handle for sending messages and generating a handle that validates all of the received messages.

Implementations

impl NlRouter

pub fn connect( proto: NlFamily, pid: Option<u32>, groups: Groups, ) -> Result<(Self, NlRouterReceiverHandle<u16, Genlmsghdr<u8, u16, NoUserHeader>>), RouterError<u16, Buffer>>

Equivalent of socket and bind calls.

pub fn add_mcast_membership( &self, groups: Groups, ) -> Result<(), RouterError<u16, Buffer>>

Join multicast groups for a socket.

pub fn drop_mcast_membership( &self, groups: Groups, ) -> Result<(), RouterError<u16, Buffer>>

Leave multicast groups for a socket.

pub fn list_mcast_membership( &self, ) -> Result<NetlinkBitArray, RouterError<u16, Buffer>>

List joined groups for a socket.

pub fn enable_ext_ack( &self, enable: bool, ) -> Result<(), RouterError<u16, Buffer>>

If true is passed in, enable extended ACKs for this socket. If false is passed in, disable extended ACKs for this socket.

pub fn get_ext_ack_enabled(&self) -> Result<bool, RouterError<u16, Buffer>>

Return true if an extended ACK is enabled for this socket.

pub fn enable_strict_checking( &self, enable: bool, ) -> Result<(), RouterError<u16, Buffer>>

If true is passed in, enable strict checking for this socket. If false is passed in, disable strict checking for for this socket. Only supported by NlFamily::Route sockets. Requires Linux >= 4.20.

pub fn get_strict_checking_enabled( &self, ) -> Result<bool, RouterError<u16, Buffer>>

Return true if strict checking is enabled for this socket. Only supported by NlFamily::Route sockets. Requires Linux >= 4.20.

pub fn pid(&self) -> u32

Get the PID for the current socket.

pub fn send<ST, SP, RT, RP>( &self, nl_type: ST, nl_flags: NlmF, nl_payload: NlPayload<ST, SP>, ) -> Result<NlRouterReceiverHandle<RT, RP>, RouterError<ST, SP>>
where ST: NlType, SP: Size + ToBytes,

Send a message and return a handle for receiving responses from this message.

pub fn resolve_genl_family( &self, family_name: &str, ) -> Result<u16, RouterError<GenlId, Genlmsghdr<CtrlCmd, CtrlAttr>>>

Convenience function for resolving a str containing the generic netlink family name to a numeric generic netlink ID.

pub fn resolve_nl_mcast_group( &self, family_name: &str, mcast_name: &str, ) -> Result<u32, RouterError<GenlId, Genlmsghdr<CtrlCmd, CtrlAttr>>>

Convenience function for resolving a str containing the multicast group name to a numeric multicast group ID.

pub fn lookup_id( &self, id: u32, ) -> Result<(String, String), RouterError<GenlId, Genlmsghdr<CtrlCmd, CtrlAttr>>>

Look up netlink family and multicast group name by ID.