Struct neli::rtnl::Ifinfomsg

pub struct Ifinfomsg {
    pub ifi_family: RtAddrFamily,
    pub ifi_type: Arphrd,
    pub ifi_index: c_int,
    pub ifi_flags: IffFlags,
    pub ifi_change: IffFlags,
    pub rtattrs: RtBuffer<Ifla, Buffer>,
    /* private fields */
}

Struct representing interface information messages

Fields
ifi_family: RtAddrFamily

Interface address family

ifi_type: Arphrd

Interface type

ifi_index: c_int

Interface index

ifi_flags: IffFlags

Interface flags

ifi_change: IffFlags

Interface change mask

rtattrs: RtBuffer<Ifla, Buffer>

Payload of Rtattrs

Implementations

impl Ifinfomsg

pub fn new(
ifi_family: RtAddrFamily,
ifi_type: Arphrd,
ifi_index: c_int,
ifi_flags: IffFlags,
ifi_change: IffFlags,
rtattrs: RtBuffer<Ifla, Buffer>
) -> Self

Create a fully initialized interface info struct

pub fn up(
ifi_family: RtAddrFamily,
ifi_type: Arphrd,
ifi_index: c_int,
rtattrs: RtBuffer<Ifla, Buffer>
) -> Self

Set the link with the given index up (equivalent to ip link set dev DEV up)

pub fn down(
ifi_family: RtAddrFamily,
ifi_type: Arphrd,
ifi_index: c_int,
rtattrs: RtBuffer<Ifla, Buffer>
) -> Self

Set the link with the given index down (equivalent to ip link set dev DEV down)