neli::rtnl

Struct Rtattr

pub struct Rtattr<T, P> {
    pub rta_len: c_ushort,
    pub rta_type: T,
    pub rta_payload: P,
}

Struct representing route netlink attributes

Fields
rta_len: c_ushort

Length of the attribute

rta_type: T

Type of the attribute

rta_payload: P

Payload of the attribute

Implementations

impl<T> Rtattr<T, Buffer>
where T: RtaType,

pub fn new<P>( _: Option<u16>, rta_type: T, rta_payload: P, ) -> Result<Self, SerError>
where P: Size + ToBytes,

Create a new Rtattr.

pub fn add_nested_attribute<TT, P>( &mut self, attr: &Rtattr<TT, P>, ) -> Result<(), SerError>
where TT: RtaType, P: ToBytes,

Add a nested attribute to the end of the payload.

pub fn get_attr_handle<R>( &self, ) -> Result<AttrHandle<'_, RtBuffer<R, Buffer>, Rtattr<R, Buffer>>, DeError>
where R: RtaType,

Return an AttrHandle for attributes nested in the given attribute payload.

pub fn get_attr_handle_mut<R>( &mut self, ) -> Result<AttrHandleMut<'_, RtBuffer<R, Buffer>, Rtattr<R, Buffer>>, DeError>
where R: RtaType,

Return an AttrHandleMut for attributes nested in the given attribute payload.