Struct neli::socket::NlSocket

pub struct NlSocket { /* fields omitted */ }

Handle for the socket file descriptor

Methods

impl NlSocket

pub fn new(proto: NlFamily) -> Result<Self, Error>

Wrapper around socket() syscall filling in the netlink-specific information

pub fn bind(
&mut self,
pid: Option<u32>,
groups: Option<u32>
) -> Result<(), Error>

Use this function to bind to a netlink ID and subscribe to groups. See netlink(7) man pages for more information on netlink IDs and groups.

pub fn send(&mut self, buf: &[u8], flags: i32) -> Result<isize, Error>

Send message encoded as byte slice to the netlink ID specified in the netlink header (nl::nlhdr::NlHdr).

pub fn recv(&mut self, len: Option<usize>, flags: i32) -> Result<Vec<u8>, Error>

Receive message encoded as byte slice from the netlink socket.

pub fn connect(
proto: NlFamily,
pid: Option<u32>,
groups: Option<u32>
) -> Result<Self, Error>

Equivalent of socket and bind calls.

pub fn sendmsg<I: Nl, T: Nl>(
&mut self,
msg: NlHdr<I, T>,
flags: i32
) -> Result<isize, NlError>

Serialize and send Rust NlMsg type

pub fn recvmsg<I: Nl, T: Nl>(
&mut self,
len: Option<usize>,
flags: i32
) -> Result<NlHdr<I, T>, NlError>

Receive and deserialize Rust NlMsg type