Struct neli::socket::NlSocketHandle

pub struct NlSocketHandle { /* private fields */ }

Higher level handle for socket operations.

Implementations

impl NlSocketHandle

pub fn new(proto: NlFamily) -> Result<Self, Error>

Wrapper around socket() syscall filling in the netlink-specific information

pub fn connect(
proto: NlFamily,
pid: Option<u32>,
groups: &[u32]
) -> Result<Self, Error>

Equivalent of socket and bind calls.

pub fn block(&self) -> Result<(), Error>

Set underlying socket file descriptor to be blocking.

pub fn nonblock(&self) -> Result<(), Error>

Set underlying socket file descriptor to be non blocking.

pub fn is_blocking(&self) -> Result<bool, Error>

Determines if underlying file descriptor is blocking.

pub fn bind(&self, pid: Option<u32>, groups: &[u32]) -> Result<(), Error>

Use this function to bind to a netlink ID and subscribe to groups. See netlink(7) man pages for more information on netlink IDs and groups.

pub fn add_mcast_membership(&self, groups: &[u32]) -> Result<(), Error>

Join multicast groups for a socket.

pub fn drop_mcast_membership(&self, groups: &[u32]) -> Result<(), Error>

Leave multicast groups for a socket.

pub fn list_mcast_membership(&self) -> Result<NetlinkBitArray, Error>

List joined groups for a socket.

pub fn resolve_genl_family(&mut self, family_name: &str) -> Result<u16, NlError>

Convenience function for resolving a str containing the generic netlink family name to a numeric generic netlink ID.

pub fn resolve_nl_mcast_group(
&mut self,
family_name: &str,
mcast_name: &str
) -> Result<u32, NlError>

Convenience function for resolving a str containing the multicast group name to a numeric multicast group ID.

pub fn lookup_id(&mut self, id: u32) -> Result<(String, String), NlError>

Look up netlink family and multicast group name by ID.

pub fn send<T, P>(&mut self, msg: Nlmsghdr<T, P>) -> Result<(), NlError>where
T: Nl + NlType + Debug,
P: Nl + Debug,

Convenience function to send an Nlmsghdr struct

pub fn recv<T, P>(&mut self) -> Result<Option<Nlmsghdr<T, P>>, NlError>where
T: Nl + NlType + Debug,
P: Nl + Debug,

Convenience function to read a stream of Nlmsghdr structs one by one. Use NlSocketHandle::iter or [NlSockHandle::iter2] instead for easy iteration over returned packets.

Returns None only in non-blocking contexts if no message can be immediately returned or if the socket has been closed.

pub fn recv_all<T, P>(&mut self) -> Result<NlBuffer<T, P>, NlError>where
T: Nl + NlType + Debug,
P: Nl + Debug,

Parse all Nlmsghdr structs sent in one network packet and return them all in a list.

Failure to parse any packet will cause the entire operation to fail. If an error is detected at the application level, this method will discard any non-error Nlmsghdr structs and only return the error. This method checks for ACKs. For a more granular approach, use either NlSocketHandle::recv or NlSocketHandle::iter.

pub fn iter<P>(
&mut self,
iter_indefinitely: bool
) -> NlMessageIter<'_, NlTypeWrapper, P>where
P: Nl + Debug,

Return an iterator object

The argument iterate_indefinitely is documented in more detail in NlMessageIter

pub fn iter2<T, P>(&mut self, iter_indefinitely: bool) -> NlMessageIter<'_, T, P>where
T: NlType + Debug,
P: Nl + Debug,

Return an iterator object.

This method allows more flexibility than NlSocketHandle::iter() and allows specifying what the type of the netlink packets should be as well as the payload.

The argument iterate_indefinitely is documented in more detail in NlMessageIter