neli::socket::synchronous

Struct NlSocketHandle

pub struct NlSocketHandle { /* private fields */ }

Higher level handle for socket operations.

Implementations

impl NlSocketHandle

pub fn connect( proto: NlFamily, pid: Option<u32>, groups: Groups, ) -> Result<Self, SocketError>

Equivalent of socket and bind calls.

pub fn add_mcast_membership(&self, groups: Groups) -> Result<(), SocketError>

Join multicast groups for a socket.

pub fn drop_mcast_membership(&self, groups: Groups) -> Result<(), SocketError>

Leave multicast groups for a socket.

pub fn list_mcast_membership(&self) -> Result<NetlinkBitArray, SocketError>

List joined groups for a socket.

pub fn pid(&self) -> u32

Get the PID for the current socket.

pub fn send<T, P>(&self, msg: &Nlmsghdr<T, P>) -> Result<(), SocketError>
where T: NlType + Debug, P: Size + ToBytes + Debug,

Convenience function to send an Nlmsghdr struct

pub fn recv<T, P>( &self, ) -> Result<(NlBufferIter<T, P, BufferPoolGuard<'_>>, Groups), SocketError>
where T: NlType + Debug, P: Size + FromBytesWithInput<Input = usize> + Debug,

Convenience function to read a stream of Nlmsghdr structs one by one using an iterator.

Returns None when the stream of messages has been completely processed in the current buffer resulting from a single NlSocket::recv call.

See NlBufferIter for more detailed information.

pub fn recv_all<T, P>(&self) -> Result<(NlBuffer<T, P>, Groups), SocketError>
where T: NlType + Debug, P: Size + FromBytesWithInput<Input = usize> + Debug,

Parse all Nlmsghdr structs sent in one network packet and return them all in a list.

Failure to parse any packet will cause the entire operation to fail. If an error is detected at the application level, this method will discard any non-error Nlmsghdr structs and only return the error. For a more granular approach, use NlSocketHandle::recv.

pub fn set_recv_buffer_size(&self, size: usize) -> Result<(), SocketError>

Set the size of the receive buffer for the socket.

This can be useful when communicating with a service that sends a high volume of messages (especially multicast), and your application cannot process them fast enough, leading to the kernel dropping messages. A larger buffer may help mitigate this.

The value passed is a hint to the kernel to set the size of the receive buffer. The kernel will double the value provided to account for bookkeeping overhead. The doubled value is capped by the value in /proc/sys/net/core/rmem_max.

The default value is /proc/sys/net/core/rmem_default

See socket(7) documentation for SO_RCVBUF for more information.

pub fn enable_ext_ack(&self, enable: bool) -> Result<(), SocketError>

If true is passed in, enable extended ACKs for this socket. If false is passed in, disable extended ACKs for this socket.

pub fn get_ext_ack_enabled(&self) -> Result<bool, SocketError>

Return true if an extended ACK is enabled for this socket.

pub fn enable_strict_checking(&self, enable: bool) -> Result<(), SocketError>

If true is passed in, enable strict checking for this socket. If false is passed in, disable strict checking for for this socket. Only supported by NlFamily::Route sockets. Requires Linux >= 4.20.

pub fn get_strict_checking_enabled(&self) -> Result<bool, SocketError>

Return true if strict checking is enabled for this socket. Only supported by NlFamily::Route sockets. Requires Linux >= 4.20.