neli

Trait FromBytes

pub trait FromBytes<'a>: Sized + Debug {
    // Required method
    fn from_bytes(buffer: &mut Cursor<&'a [u8]>) -> Result<Self, DeError>;

    // Provided method
    fn strip(buffer: &mut Cursor<&'a [u8]>) -> Result<(), DeError> { ... }
}

A trait defining how to convert from a byte buffer to a netlink data structure.

Required Methods

fn from_bytes(buffer: &mut Cursor<&'a [u8]>) -> Result<Self, DeError>

Takes a byte buffer and returns the deserialized data structure.

Provided Methods

fn strip(buffer: &mut Cursor<&'a [u8]>) -> Result<(), DeError>

Strip padding from a netlink message.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.