Trait neli::Nl

pub trait Nl: Sized {
    fn serialize(&self, m: SerBuffer<'_>) -> Result<(), SerError>;
    fn deserialize(m: DeBuffer<'_>) -> Result<Self, DeError>;
    fn type_size() -> Option<usize>;
    fn size(&self) -> usize;

    fn type_asize() -> Option<usize> { ... }
    fn asize(&self) -> usize { ... }
    fn pad(&self, mem: SerBuffer<'_>) -> Result<(), SerError> { ... }
}

Trait defining basic actions required for netlink communication. Implementations for basic and neli’s types are provided (see below). Create new implementations if you have to work with a Netlink API that uses values of more unusual types.

Required Methods

fn serialize(&self, m: SerBuffer<'_>) -> Result<(), SerError>

Serialization method

fn deserialize(m: DeBuffer<'_>) -> Result<Self, DeError>

Deserialization method

fn type_size() -> Option<usize>

The size of the binary representation of a type not aligned to 4-byte boundary size

fn size(&self) -> usize

The size of the binary representation of an existing value not aligned to 4-byte boundary size

Provided Methods

fn type_asize() -> Option<usize>

The size of the binary representation of a type not aligned to 4-byte boundary size

fn asize(&self) -> usize

The size of the binary representation of an existing value aligned to 4-byte boundary size

fn pad(&self, mem: SerBuffer<'_>) -> Result<(), SerError>

Pad the data serialized data structure to alignment