neli

Trait Size

pub trait Size {
    // Required method
    fn unpadded_size(&self) -> usize;

    // Provided method
    fn padded_size(&self) -> usize { ... }
}

A trait defining methods that apply to all netlink data structures related to sizing of data types.

Required Methods

fn unpadded_size(&self) -> usize

Size of the unpadded data structure. This will usually only be unaligned for variable length types like strings or byte buffers.

Provided Methods

fn padded_size(&self) -> usize

Get the size of the payload and align it to the required netlink byte alignment.