Module neli::types

Module containing various types used across the various netlink structures used in neli.

Design decisions

These structures are new types rather than type aliases in most cases to allow the internal representation to change without resulting in a breaking change.

Structs

Buffer

A buffer of bytes.

FlagBuffer

A buffer of flag constants.

GenlBuffer

A buffer of generic netlink attributes.

Iter

Immutable slice iterator

IterMut

Mutable slice iterator.

NlBuffer

A buffer of netlink messages.

PhantomData

Zero-sized type used to mark things that “act like” they own a T.

Range

A (half-open) range bounded inclusively below and exclusively above (start..end).

Ref

Wraps a borrowed reference to a value in a RefCell box. A wrapper type for an immutably borrowed value from a RefCell<T>.

RefCell

A mutable memory location with dynamically checked borrow rules

RefMut

A wrapper type for a mutably borrowed value from a RefCell<T>.

RtBuffer

A buffer of rtnetlink attributes.

SockBuffer

A buffer to hold data read from sockets

SockBufferRef

An immutable reference to the socket buffer.

SockBufferRefMut

A mutable reference to the socket buffer.

Traits

FromIterator

Conversion from an Iterator.

Type Definitions

DeBuffer

Type alias for a buffer to deserialize from.

SerBuffer

Type alias for a buffer to serialize into.