Struct neli::types::NlBuffer

pub struct NlBuffer<T, P>(_);

A buffer of netlink messages.

Implementations

impl<T, P> NlBuffer<T, P>

pub fn new() -> Self

Create a new buffer of netlink messages.

pub fn push(&mut self, msg: Nlmsghdr<T, P>)

Add a new netlink message to the end of the buffer.

pub fn pop(&mut self) -> Option<Nlmsghdr<T, P>>

Get a netlink message from the end of the buffer.

pub fn iter(&self) -> Iter<'_, Nlmsghdr<T, P>>

Return an iterator over immutable references to the elements in the buffer.

pub fn iter_mut(&mut self) -> IterMut<'_, Nlmsghdr<T, P>>

Return an iterator over mutable references to the elements in the buffer.

pub fn len(&self) -> usize

Returns the number of elements in the buffer.

pub fn is_empty(&self) -> bool

Returns whether the number of elements in the buffer is 0.