neli::types
Struct RtBuffer

pub struct RtBuffer<T, P>(/* private fields */);

A buffer of rtnetlink attributes.

Implementations

impl<T> RtBuffer<T, Buffer>

pub fn get_attr_handle(&self) -> AttrHandle<'_, Self, Rtattr<T, Buffer>>

Get a data structure with an immutable reference to the underlying Rtattrs.

pub fn get_attr_handle_mut( &mut self, ) -> AttrHandleMut<'_, Self, Rtattr<T, Buffer>>

Get a data structure with a mutable reference to the underlying Rtattrs.

impl<T, P> RtBuffer<T, P>

pub fn new() -> Self

Create a new buffer of routing netlink attributes.

pub fn push(&mut self, attr: Rtattr<T, P>)

Add a new routing netlink attribute to the end of the buffer.

pub fn pop(&mut self) -> Option<Rtattr<T, P>>

Get a routing netlink attribute from the end of the buffer.

pub fn iter(&self) -> Iter<'_, Rtattr<T, P>>

Return an iterator over immutable references to the elements in the buffer.

pub fn iter_mut(&mut self) -> IterMut<'_, Rtattr<T, P>>

Return an iterator over mutable references to the elements in the buffer.

pub fn len(&self) -> usize

Returns the number of elements in the buffer.

pub fn is_empty(&self) -> bool

Returns whether the number of elements in the buffer is 0.