neli::utils
Struct NetlinkBitArray

pub struct NetlinkBitArray(/* private fields */);

A bit array meant to be compatible with the bit array returned by the NETLINK_LIST_MEMBERSHIPS socket operation on netlink sockets.

Implementations

impl NetlinkBitArray

bittest/bitset instrinsics are not stable in Rust so this needs to be implemented this way.

pub fn new(bit_len: usize) -> Self

Create a new bit array.

This method will round bit_len up to the nearest multiple of size_of::<u32>().

pub fn resize_bits(&mut self, bit_len: usize)

Resize the underlying vector to have enough space for the nearest multiple of size_of::<u32>() rounded up.

pub fn resize(&mut self, bytes: usize)

Resize the underlying vector to have enough space for the nearest multiple of size_of::<BitArrayType>().

pub fn is_set(&self, n: usize) -> bool

Returns true if the nth bit is set.

pub fn set(&mut self, n: usize)

Set the nth bit.

pub fn to_vec(&self) -> Vec<u32>

Get a vector representation of all of the bit positions set to 1 in this bit array.

Example

use neli::utils::NetlinkBitArray;

let mut array = NetlinkBitArray::new(24);
array.set(4);
array.set(7);
array.set(23);
assert_eq!(array.to_vec(), vec![4, 7, 23]);

pub fn len_bits(&self) -> usize

Return the number of bits that can be contained in this bit array.

pub fn len(&self) -> usize

Return the length in bytes for this bit array.