neli::consts

Module netfilter

Constants related to netfilter netlink integration Constants for netfilter related protocols

Note that this doesn’t cover everything yet, both the list of types and variants in enums will be added over time.

Enums

LogCfgCmdWrapper

Wrapper that is valid anywhere that accepts a value implementing the LogCfgCmd trait

LogCmd

Command value for the NfLogCfg::Cmd.

LogCopyMode

Copy mode of the logged packets.

NetfilterMsg

Messages related to the netfilter netlink protocols.

NfLogAttr

Attributes inside a netfilter log packet message.

NfLogCfg

Configuration attributes for netfilter logging.

Traits

LogCfgCmd

Parameters for the NfLogCfg::Cmd.