neli

Module err

This is the module that contains the error types used in neli

There are five main types:

Nlmsgerr - an application error returned from netlink as a packet.

RouterError - errors returned by NlRouter.

SocketError - errors returned by NlSocketHandle.

DeError - error while deserializing

SerError - error while serializing

Design decisions

All errors implement std::error::Error in an attempt to allow them to be used in conjunction with Result for easier error management even at the protocol error level.

Structs

MsgError

Arbitrary error message.

Nlmsgerr

Struct representing netlink packets containing errors

NlmsgerrBuilder

Builder for Nlmsgerr.

NlmsghdrAck

A special struct that represents the contents of an ACK returned at the application level.

NlmsghdrAckBuilder

Builder for NlmsghdrAck.

NlmsghdrErr

A special struct that represents the contents of an error returned at the application level.

NlmsghdrErrBuilder

Builder for NlmsghdrErr.

Enums

BuilderError

DeError

Deserialization error

NlmsgerrBuilderError

Error type for NlmsgerrBuilder

NlmsghdrAckBuilderError

Error type for NlmsghdrAckBuilder

NlmsghdrErrBuilderError

Error type for NlmsghdrErrBuilder

RouterError

Sendable, clonable error that can be sent across channels in the router infrastructure to provide typed errors to all receivers indicating what went wrong.

SerError

Serialization error

SocketError

General netlink error

Utf8

String or str UTF error.