neli

Module genl

This module contains generic netlink parsing data structures. This is all handled by the Genlmsghdr header struct which contains all of the information needed for the generic netlink layer.

Design decisions

The generic netlink attrs field has been changed to a GenlBuffer of Nlattrs instead of the original Vec<u8> to allow simpler parsing at the top level when one Nlattr structure is not nested within another, a use case that is instead handled using AttrHandle.

Structs

AttrType

The infomation packed into nla_type field of nlattr for the C data structure.

AttrTypeBuilder

Builder for AttrType.

Genlmsghdr

Struct representing generic netlink header and payload

GenlmsghdrBuilder

Builder for Genlmsghdr.

Nlattr

Struct representing netlink attributes and payloads

NlattrBuilder

Builder for Nlattr.

NoUserHeader

Struct indicating that no user header is in the generic netlink packet.

Enums

AttrTypeBuilderError

Error type for AttrTypeBuilder

GenlmsghdrBuilderError

Error type for GenlmsghdrBuilder

NlattrBuilderError

Error type for NlattrBuilder

Type Aliases

GenlAttrHandle

Type representing a generic netlink attribute handle.