neli

Module rtnl

This module provides an implementation of routing netlink structures and the routing attributes that are at the end of most routing netlink responses.

Design decisions

This module is based very heavily on the information in man 7 rtnetlink so it is mainly a series of structs organized in a style similar to the rest of the library.

Structs

Ifaddrmsg

Struct representing interface address messages

IfaddrmsgBuilder

Builder for Ifaddrmsg.

Ifinfomsg

Struct representing interface information messages

IfinfomsgBuilder

Builder for Ifinfomsg.

NdaCacheinfo

Struct representing ARP cache info

NdaCacheinfoBuilder

Builder for NdaCacheinfo.

Ndmsg

Represents an ARP (neighbor table) entry

NdmsgBuilder

Builder for Ndmsg.

Rtattr

Struct representing route netlink attributes

RtattrBuilder

Builder for Rtattr.

Rtgenmsg

General form of address family dependent message. Used for requesting things from rtnetlink.

RtgenmsgBuilder

Builder for Rtgenmsg.

Rtmsg

Route message

RtmsgBuilder

Builder for Rtmsg.

Tcmsg

Message in response to queuing discipline operations

TcmsgBuilder

Builder for Tcmsg.

Enums

IfaddrmsgBuilderError

Error type for IfaddrmsgBuilder

IfinfomsgBuilderError

Error type for IfinfomsgBuilder

NdaCacheinfoBuilderError

Error type for NdaCacheinfoBuilder

NdmsgBuilderError

Error type for NdmsgBuilder

RtattrBuilderError

Error type for RtattrBuilder

RtgenmsgBuilderError

Error type for RtgenmsgBuilder

RtmsgBuilderError

Error type for RtmsgBuilder

TcmsgBuilderError

Error type for TcmsgBuilder

Type Aliases

RtAttrHandle

Represents a routing netlink attribute handle.