neli

Module rtnl

This module provides an implementation of routing netlink structures and the routing attributes that are at the end of most routing netlink responses.

Design decisions

This module is based very heavily on the information in man 7 rtnetlink so it is mainly a series of structs organized in a style similar to the rest of the library.

Structs

Ifaddrmsg

Struct representing interface address messages

Ifinfomsg

Struct representing interface information messages

NdaCacheinfo

Struct representing ARP cache info

Ndmsg

Represents an ARP (neighbor table) entry

Rtattr

Struct representing route netlink attributes

Rtgenmsg

General form of address family dependent message. Used for requesting things from rtnetlink.

Rtmsg

Route message

Tcmsg

Message in response to queuing discipline operations