neli::socket
Struct NlSocket

pub struct NlSocket { /* private fields */ }

Low level access to a netlink socket.

Implementations

impl NlSocket

pub fn new(proto: NlFamily) -> Result<Self, Error>

Wrapper around socket() syscall filling in the netlink-specific information.

pub fn connect( proto: NlFamily, pid: Option<u32>, groups: &[u32], ) -> Result<Self, Error>

Equivalent of socket and bind calls.

pub fn block(&self) -> Result<(), Error>

Set underlying socket file descriptor to be blocking.

pub fn nonblock(&self) -> Result<(), Error>

Set underlying socket file descriptor to be non blocking.

pub fn is_blocking(&self) -> Result<bool, Error>

Determines if underlying file descriptor is blocking.

pub fn bind(&self, pid: Option<u32>, groups: &[u32]) -> Result<(), Error>

Use this function to bind to a netlink ID and subscribe to groups. See netlink(7) man pages for more information on netlink IDs and groups.

pub fn add_mcast_membership(&self, groups: &[u32]) -> Result<(), Error>

Join multicast groups for a socket.

pub fn drop_mcast_membership(&self, groups: &[u32]) -> Result<(), Error>

Leave multicast groups for a socket.

pub fn list_mcast_membership(&self) -> Result<NetlinkBitArray, Error>

List joined groups for a socket.

pub fn send<B>(&self, buf: B, flags: i32) -> Result<size_t, Error>
where B: AsRef<[u8]>,

Send message encoded as byte slice to the netlink ID specified in the netlink header Nlmsghdr

pub fn recv<B>(&self, buf: B, flags: i32) -> Result<size_t, Error>
where B: AsMut<[u8]>,

Receive message encoded as byte slice from the netlink socket.

pub fn pid(&self) -> Result<u32, Error>

Get the PID for this socket.