Trait neli::Nl

pub trait Nl: Sized + Default {
    fn serialize(&mut self, _: &mut NlSerState) -> Result<(), SerError>;

    fn deserialize(state: &mut NlDeState) -> Result<Self, DeError>;

    fn size(&self) -> usize;

    fn asize(&self) -> usize { ... }
}

Trait defining basic actions required for netlink communication

Required Methods

fn serialize(&mut self, _: &mut NlSerState) -> Result<(), SerError>

Serialization method

fn deserialize(state: &mut NlDeState) -> Result<Self, DeError>

Deserialization method

fn size(&self) -> usize

The size of the binary representation of a struct - not aligned to word size

Provided Methods

fn asize(&self) -> usize

The size of the binary representation of a struct - aligned to word size